# Optimize memory allocation when rendering partials