--- a/lib/jbuilder/jbuilder_template.rb
+++ b/lib/jbuilder/jbuilder_template.rb
@@ -244,7 +244,7 @@ class JbuilderTemplate < Jbuilder
       _scope{ _render_partial_with_options options.merge(locals: locals) }
     end
 
-    set! name, value
+    _set_value name, value
   end
 
   def _render_explicit_partial(name_or_options, locals = {})