--- a/lib/jbuilder/jbuilder_template.rb
+++ b/lib/jbuilder/jbuilder_template.rb
@@ -133,7 +133,7 @@ class JbuilderTemplate < Jbuilder
     options = args.first
 
     if args.one? && _partial_options?(options)
-      _set_inline_partial name, object, options
+      _set_inline_partial name, object, options.dup
     else
       super
     end